+++ modules/core/src/test/java/org/locationtech/jts/geom/impl/PackedCoordinateSequenceTest.java
@@ -25,7 +25,6 @@ import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.geom.Polygon;
 
 
-
 /**
  * Test {@link PackedCoordinateSequence}
  * @version 1.7
@@ -40,14 +39,14 @@ public class PackedCoordinateSequenceTest
 
   @Test
   public void testDouble() {
-    checkAll( PackedCoordinateSequenceFactory.DOUBLE_FACTORY );
+    checkAll(PackedCoordinateSequenceFactory.DOUBLE_FACTORY);
   }
 
   @Test
   public void testFloat() {
-    checkAll( PackedCoordinateSequenceFactory.FLOAT_FACTORY) ;
+    checkAll(PackedCoordinateSequenceFactory.FLOAT_FACTORY);
   }
-  
+
   public void checkAll(CoordinateSequenceFactory factory)
   {
     checkDim2(1, factory);
@@ -58,116 +57,116 @@ public class PackedCoordinateSequenceTest
     checkDim4(factory);
     checkDimInvalid(factory);
   }
-  
+
   public void checkDim2(int size, CoordinateSequenceFactory factory)
   {
     CoordinateSequence seq = factory.create(size, 2);
     initProgression(seq);
-    
+
     assertEquals(2, seq.getDimension(), "Dimension should be 2");
     assertTrue(!seq.hasZ(), "Z should not be present");
     assertTrue(!seq.hasM(), "M should not be present");
-    
+
     int indexLast = size - 1;
     double valLast = indexLast;
-    
+
     Coordinate coord = seq.getCoordinate(indexLast);
-    assertTrue( coord instanceof CoordinateXY);
-    assertEquals( valLast, coord.getX());
-    assertEquals( valLast, coord.getY());
-    
+    assertTrue(coord instanceof CoordinateXY);
+    assertEquals(valLast, coord.getX());
+    assertEquals(valLast, coord.getY());
+
     Coordinate[] array = seq.toCoordinateArray();
     assertEquals(coord, array[indexLast]);
     assertTrue(coord != array[indexLast]);
-    assertTrue(isEqual(seq,array));
-    
+    assertTrue(isEqual(seq, array));
+
     CoordinateSequence copy = factory.create(array);
-    assertTrue(isEqual(copy,array));
-    
+    assertTrue(isEqual(copy, array));
+
     CoordinateSequence copy2 = factory.create(seq);
-    assertTrue(isEqual(copy2,array));
+    assertTrue(isEqual(copy2, array));
   }
-  
+
   public void checkDim3(CoordinateSequenceFactory factory)
   {
     CoordinateSequence seq = factory.create(5, 3);
     initProgression(seq);
-    
+
     assertEquals(3, seq.getDimension(), "Dimension should be 3");
     assertTrue(seq.hasZ(), "Z should be present");
     assertTrue(!seq.hasM(), "M should not be present");
-    
+
     Coordinate coord = seq.getCoordinate(4);
-    assertTrue( coord.getClass() == Coordinate.class);
-    assertEquals( 4.0, coord.getX());
-    assertEquals( 4.0, coord.getY());
-    assertEquals( 4.0, coord.getZ());
-    
+    assertTrue(coord.getClass() == Coordinate.class);
+    assertEquals(4.0, coord.getX());
+    assertEquals(4.0, coord.getY());
+    assertEquals(4.0, coord.getZ());
+
     Coordinate[] array = seq.toCoordinateArray();
     assertEquals(coord, array[4]);
     assertTrue(coord != array[4]);
     assertTrue(isEqual(seq, array));
-    
+
     CoordinateSequence copy = factory.create(array);
     assertTrue(isEqual(copy, array));
-    
+
     CoordinateSequence copy2 = factory.create(seq);
     assertTrue(isEqual(copy2, array));
   }
-    
+
   public void checkDim3_M1(CoordinateSequenceFactory factory)
-  { 
+  {
     CoordinateSequence seq = factory.create(5, 3, 1);
-    initProgression(seq);   
-    
+    initProgression(seq);
+
     assertEquals(3, seq.getDimension(), "Dimension should be 3");
     assertTrue(!seq.hasZ(), "Z should not be present");
     assertTrue(seq.hasM(), "M should be present");
-    
+
     Coordinate coord = seq.getCoordinate(4);
-    assertTrue( coord instanceof CoordinateXYM);
-    assertEquals( 4.0, coord.getX());
-    assertEquals( 4.0, coord.getY());
-    assertEquals( 4.0, coord.getM());
-    
+    assertTrue(coord instanceof CoordinateXYM);
+    assertEquals(4.0, coord.getX());
+    assertEquals(4.0, coord.getY());
+    assertEquals(4.0, coord.getM());
+
     Coordinate[] array = seq.toCoordinateArray();
     assertEquals(coord, array[4]);
     assertTrue(coord != array[4]);
-    assertTrue(isEqual(seq,array));
-    
+    assertTrue(isEqual(seq, array));
+
     CoordinateSequence copy = factory.create(array);
-    assertTrue(isEqual(copy,array));
-    
+    assertTrue(isEqual(copy, array));
+
     CoordinateSequence copy2 = factory.create(seq);
-    assertTrue(isEqual(copy2,array));
+    assertTrue(isEqual(copy2, array));
   }
-  
+
   public void checkDim4_M1(CoordinateSequenceFactory factory)
   {
     CoordinateSequence seq = factory.create(5, 4, 1);
     initProgression(seq);
-    
+
     assertEquals(4, seq.getDimension(), "Dimension should be 4");
     assertTrue(seq.hasZ(), "Z should be present");
     assertTrue(seq.hasM(), "M should be present");
-    
+
     Coordinate coord = seq.getCoordinate(4);
-    assertTrue( coord instanceof CoordinateXYZM);
-    assertEquals( 4.0, coord.getX());
-    assertEquals( 4.0, coord.getY());
-    assertEquals( 4.0, coord.getZ());
-    assertEquals( 4.0, coord.getM());
-    
+    assertTrue(coord instanceof CoordinateXYZM);
+    assertEquals(4.0, coord.getX());
+    assertEquals(4.0, coord.getY());
+    assertEquals(4.0, coord.getZ());
+    assertEquals(4.0, coord.getM());
+
     Coordinate[] array = seq.toCoordinateArray();
     assertEquals(coord, array[4]);
     assertTrue(coord != array[4]);
-    assertTrue(isEqual(seq,array));
-    
+    assertTrue(isEqual(seq, array));
+
     CoordinateSequence copy = factory.create(array);
-    assertTrue(isEqual(copy,array));
-    
+    assertTrue(isEqual(copy, array));
+
     CoordinateSequence copy2 = factory.create(seq);
-    assertTrue(isEqual(copy2, array));    
+    assertTrue(isEqual(copy2, array));
   }
 
   public void checkDim4(CoordinateSequenceFactory factory)
@@ -180,19 +179,19 @@ public class PackedCoordinateSequenceTest
     assertTrue(seq.hasM(), "M should be present");
 
     Coordinate coord = seq.getCoordinate(4);
-    assertTrue( coord instanceof CoordinateXYZM);
-    assertEquals( 4.0, coord.getX());
-    assertEquals( 4.0, coord.getY());
-    assertEquals( 4.0, coord.getZ());
-    assertEquals( 4.0, coord.getM());
+    assertTrue(coord instanceof CoordinateXYZM);
+    assertEquals(4.0, coord.getX());
+    assertEquals(4.0, coord.getY());
+    assertEquals(4.0, coord.getZ());
+    assertEquals(4.0, coord.getM());
 
     Coordinate[] array = seq.toCoordinateArray();
     assertEquals(coord, array[4]);
     assertTrue(coord != array[4]);
-    assertTrue(isEqual(seq,array));
+    assertTrue(isEqual(seq, array));
 
     CoordinateSequence copy = factory.create(array);
-    assertTrue(isEqual(copy,array));
+    assertTrue(isEqual(copy, array));
 
     CoordinateSequence copy2 = factory.create(seq);
     assertTrue(isEqual(copy2, array));
@@ -206,34 +205,34 @@ public class PackedCoordinateSequenceTest
     GeometryFactory factoryPacked = new GeometryFactory(new PackedCoordinateSequenceFactory());
     Polygon polygonPacked = factoryPacked
         .createPolygon(
-            new Coordinate[] { 
-                 new CoordinateXY(0, 0), new CoordinateXY(10, 0), new CoordinateXY(10, 10),
-                 new CoordinateXY(0, 10), new CoordinateXY(0, 0) });
+            new Coordinate[]{
+                new CoordinateXY(0, 0), new CoordinateXY(10, 0), new CoordinateXY(10, 10),
+                new CoordinateXY(0, 10), new CoordinateXY(0, 0)});
     GeometryFactory factoryDefault = new GeometryFactory();
     Polygon polygonArray = factoryDefault.createPolygon(
-        new Coordinate[] { new CoordinateXY(5, 5),
-            new CoordinateXY(15, 5), new CoordinateXY(15, 15), new CoordinateXY(5, 15), new CoordinateXY(5, 5) });
+        new Coordinate[]{new CoordinateXY(5, 5),
+            new CoordinateXY(15, 5), new CoordinateXY(15, 15), new CoordinateXY(5, 15), new CoordinateXY(5, 5)});
 
     polygonArray.intersection(polygonPacked);
-    
+
     // this fails as of 2019-June-7
     polygonPacked.intersection(polygonArray);
   }
-  
+
   public void checkDimInvalid(CoordinateSequenceFactory factory)
   {
     try {
       CoordinateSequence seq = factory.create(5, 2, 1);
       fail("Dimension=2/Measure=1 (XM) not supported");
     } catch (IllegalArgumentException expected) {
-    }    
+    }
   }
-  
+
   private void initProgression(CoordinateSequence seq) {
-    for (int index = 0; index < seq.size(); index++) {
-       for( int ordinateIndex = 0; ordinateIndex < seq.getDimension(); ordinateIndex++) {
-         seq.setOrdinate(index, ordinateIndex, (double) index);
-       }
+    for (int index = 0;index < seq.size();index++) {
+      for (int ordinateIndex = 0;ordinateIndex < seq.getDimension();ordinateIndex++) {
+        seq.setOrdinate(index, ordinateIndex, index);
+      }
     }
   }
