+++ modules/core/src/test/java/org/locationtech/jts/io/WKBWriterTest.java
@@ -34,69 +34,69 @@ import test.jts.GeometryTestCase;
 public class WKBWriterTest extends GeometryTestCase {
   @Test
   public void testSRID() throws Exception {
-      GeometryFactory gf = new GeometryFactory();
-      Point p1 = gf.createPoint(new Coordinate(1,2));
-      p1.setSRID(1234);
-      
-      //first write out without srid set
-      WKBWriter w = new WKBWriter();
-      byte[] wkb = w.write(p1);
-      
-      //check the 3rd bit of the second byte, should be unset
-      byte b = (byte) (wkb[1] & 0x20);
-      assertEquals(0, b);
-      
-      //read geometry back in
-      WKBReader r = new WKBReader(gf);
-      Point p2 = (Point) r.read(wkb);
-      
-      assertTrue(p1.equalsExact(p2));
-      assertEquals(0, p2.getSRID());
-      
-      //not write out with srid set
-      w = new WKBWriter(2, true);
-      wkb = w.write(p1);
-      
-      //check the 3rd bit of the second byte, should be set
-      b = (byte) (wkb[1] & 0x20);
-      assertEquals(0x20, b);
-      
-      int srid = ((int) (wkb[5] & 0xff) << 24) | ( (int) (wkb[6] & 0xff) << 16)
-          | ( (int) (wkb[7] & 0xff) << 8) | (( int) (wkb[8] & 0xff) );
-     
-      assertEquals(1234, srid);
-      
-      r = new WKBReader(gf);
-      p2 = (Point) r.read(wkb);
-      
-      //read the geometry back in
-      assertTrue(p1.equalsExact(p2));
-      assertEquals(1234, p2.getSRID());
+    GeometryFactory gf = new GeometryFactory();
+    Point p1 = gf.createPoint(new Coordinate(1, 2));
+    p1.setSRID(1234);
+
+    //first write out without srid set
+    WKBWriter w = new WKBWriter();
+    byte[] wkb = w.write(p1);
+
+    //check the 3rd bit of the second byte, should be unset
+    byte b = (byte) (wkb[1] & 0x20);
+    assertEquals(0, b);
+
+    //read geometry back in
+    WKBReader r = new WKBReader(gf);
+    Point p2 = (Point) r.read(wkb);
+
+    assertTrue(p1.equalsExact(p2));
+    assertEquals(0, p2.getSRID());
+
+    //not write out with srid set
+    w = new WKBWriter(2, true);
+    wkb = w.write(p1);
+
+    //check the 3rd bit of the second byte, should be set
+    b = (byte) (wkb[1] & 0x20);
+    assertEquals(0x20, b);
+
+    int srid = ((wkb[5] & 0xff) << 24) | ((wkb[6] & 0xff) << 16)
+        | ((wkb[7] & 0xff) << 8) | (wkb[8] & 0xff);
+
+    assertEquals(1234, srid);
+
+    r = new WKBReader(gf);
+    p2 = (Point) r.read(wkb);
+
+    //read the geometry back in
+    assertTrue(p1.equalsExact(p2));
+    assertEquals(1234, p2.getSRID());
   }
 
   @Test
   public void testPointEmpty2D() {
-    checkWKB("POINT EMPTY", 2, "0101000000000000000000F87F000000000000F87F" );    
+    checkWKB("POINT EMPTY", 2, "0101000000000000000000F87F000000000000F87F");
   }
 
   @Test
   public void testPointEmpty3D() {
-    checkWKB("POINT EMPTY", 3, "0101000080000000000000F87F000000000000F87F000000000000F87F" );    
+    checkWKB("POINT EMPTY", 3, "0101000080000000000000F87F000000000000F87F000000000000F87F");
   }
 
   @Test
   public void testPolygonEmpty2DSRID() {
-    checkWKB("POLYGON EMPTY", 2, ByteOrderValues.LITTLE_ENDIAN, 4326, "0103000020E610000000000000" );    
+    checkWKB("POLYGON EMPTY", 2, ByteOrderValues.LITTLE_ENDIAN, 4326, "0103000020E610000000000000");
   }
 
   @Test
   public void testPolygonEmpty2D() {
-    checkWKB("POLYGON EMPTY", 2, "010300000000000000" );    
+    checkWKB("POLYGON EMPTY", 2, "010300000000000000");
   }
 
   @Test
   public void testPolygonEmpty3D() {
-    checkWKB("POLYGON EMPTY", 3, "010300008000000000" );    
+    checkWKB("POLYGON EMPTY", 3, "010300008000000000");
   }
 
   @Test
@@ -146,42 +146,42 @@ public class WKBWriterTest extends GeometryTestCase {
 
   @Test
   public void testWkbLineStringZM() throws ParseException {
-      LineString lineZM = new GeometryFactory().createLineString(new Coordinate[]{new CoordinateXYZM(1,2,3,4), new CoordinateXYZM(5,6,7,8)});
-      byte[] write = new WKBWriter(4).write(lineZM);
+    LineString lineZM = new GeometryFactory().createLineString(new Coordinate[]{new CoordinateXYZM(1, 2, 3, 4), new CoordinateXYZM(5, 6, 7, 8)});
+    byte[] write = new WKBWriter(4).write(lineZM);
 
-      LineString lineZMRead = (LineString) new WKBReader().read(write);
+    LineString lineZMRead = (LineString) new WKBReader().read(write);
 
-      assertEquals(lineZM, lineZMRead);
+    assertEquals(lineZM, lineZMRead);
 
-      assertEquals(1.0, lineZMRead.getPointN(0).getCoordinate().getX());
-      assertEquals(2.0, lineZMRead.getPointN(0).getCoordinate().getY());
-      assertEquals(3.0, lineZMRead.getPointN(0).getCoordinate().getZ());
-      assertEquals(4.0, lineZMRead.getPointN(0).getCoordinate().getM());
+    assertEquals(1.0, lineZMRead.getPointN(0).getCoordinate().getX());
+    assertEquals(2.0, lineZMRead.getPointN(0).getCoordinate().getY());
+    assertEquals(3.0, lineZMRead.getPointN(0).getCoordinate().getZ());
+    assertEquals(4.0, lineZMRead.getPointN(0).getCoordinate().getM());
 
-      assertEquals(5.0, lineZMRead.getPointN(1).getCoordinate().getX());
-      assertEquals(6.0, lineZMRead.getPointN(1).getCoordinate().getY());
-      assertEquals(7.0, lineZMRead.getPointN(1).getCoordinate().getZ());
-      assertEquals(8.0, lineZMRead.getPointN(1).getCoordinate().getM());
+    assertEquals(5.0, lineZMRead.getPointN(1).getCoordinate().getX());
+    assertEquals(6.0, lineZMRead.getPointN(1).getCoordinate().getY());
+    assertEquals(7.0, lineZMRead.getPointN(1).getCoordinate().getZ());
+    assertEquals(8.0, lineZMRead.getPointN(1).getCoordinate().getM());
   }
 
   void checkWKB(String wkt, int dimension, String expectedWKBHex) {
     checkWKB(wkt, dimension, ByteOrderValues.LITTLE_ENDIAN, -1, expectedWKBHex);
   }
-    
+
   void checkWKB(String wkt, int dimension, int byteOrder, int srid, String expectedWKBHex) {
     Geometry geom = read(wkt);
-    
+
     // set SRID if not -1
     boolean includeSRID = false;
     if (srid >= 0) {
       includeSRID = true;
       geom.setSRID(srid);
     }
-    
+
     WKBWriter wkbWriter = new WKBWriter(dimension, byteOrder, includeSRID);
     byte[] wkb = wkbWriter.write(geom);
     String wkbHex = WKBWriter.toHex(wkb);
-    
+
     assertEquals(expectedWKBHex, wkbHex);
   }
 }
